db_conversion/requirements.txt
pymongo==4.3.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
fire==0.5.0
paramiko==3.3.1
sshtunnel==0.4.0
